Refactor medication dispensing into smaller focused components
